requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.5
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.176
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openapi-schema-pydantic==1.2.4
operand==0.2.9
packaging==23.1
posthog==3.0.1
priority==2.0.0
protobuf==4.23.1
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
quart==0.18.4
quart-cors==0.6.0
requests==2.30.0
six==1.16.0
SQLAlchemy==2.0.15
tenacity==8.2.2
toml==0.10.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
